Introduce the KeyValueStoreView using the BTreeMap<Vec<u8>, Vec<u8>>.
