+++ linera-views/src/views/mod.rs
@@ -39,9 +39,6 @@ pub mod collection_view;
 /// The `ReentrantCollectionView` implements a map structure whose keys are ordered and the values are views with concurrent access.
 pub mod reentrant_collection_view;
 
-/// The implementation of a key-value store view.
-pub mod key_value_store_view;
-
 /// Wrapping a view to compute a hash.
 pub mod hashable_wrapper;
